Moved to a server-app file configuration, modified some routes
