+++ api/package.json
@@ -2,11 +2,11 @@
   "name": "boilerbooks-api",
   "version": "2.0.0",
   "description": "Backend API for the Purdue IEEE boilerbooks accounting software",
-  "main": "src/app.js",
+  "main": "src/server.js",
   "private": true,
   "type": "module",
   "scripts": {
-    "serve": "nodemon --exec babel-node src/app.js",
+    "serve": "nodemon --exec babel-node src/server.js",
     "lint": "eslint src/ && echo No lint errors found",
     "build": "babel --env-name build src -d dist"
   },
@@ -22,6 +22,9 @@
     "multer": "^1.4.4-lts.1",
     "mysql2": "^2.3.3",
     "nodemailer": "^6.7.8",
+    "openid-client": "^5.1.9",
+    "passport": "^0.6.0",
+    "passport-http-bearer": "^1.0.1",
     "uuid": "^9.0.0",
     "winston": "^3.8.2",
     "winston-daily-rotate-file": "^4.7.1"
