+++ api/src/routes/login.js
@@ -29,9 +29,152 @@ const bcrypt_rounds = 10;
 // ---------------------------
 
 /*
-    Logs user in
+    Creates a new user account
 */
 router.post("/", async(req, res, next) => {
+    if (req.body.fname === undefined ||
+        req.body.lname === undefined ||
+        req.body.uname === undefined ||
+        req.body.email === undefined ||
+        req.body.address === undefined ||
+        req.body.city === undefined ||
+        req.body.state === undefined ||
+        req.body.zip === undefined ||
+        req.body.pass1 === undefined ||
+        req.body.pass2 === undefined ||
+        req.body.createpin === undefined) {
+        res.status(400).send("All account details must be completed");
+        return next();
+    }
+
+    if (req.body.fname === "" ||
+        req.body.lname === "" ||
+        req.body.uname === "" ||
+        req.body.email === "" ||
+        req.body.address === "" ||
+        req.body.city === "" ||
+        req.body.state === "" ||
+        req.body.zip === "" ||
+        req.body.pass1 === "" ||
+        req.body.pass2 === "" ||
+        req.body.createpin === "") {
+        res.status(400).send("All account details must be completed");
+        return next();
+    }
+
+    if (req.body.fname.length > 100) {
+        res.status(400).send("First Name is too long");
+        return next();
+    }
+    if (req.body.lname.length > 100) {
+        res.status(400).send("Last Name is too long");
+        return next();
+    }
+    if (req.body.email.length > 200) {
+        res.status(400).send("Email is too long");
+        return next();
+    }
+    if (req.body.address.length > 200) {
+        res.status(400).send("Address is too long");
+        return next();
+    }
+    if (req.body.city.length > 100) {
+        res.status(400).send("City is too long");
+        return next();
+    }
+
+    if (req.body.state.length !== 2) {
+        res.status(400).send("State must be a 2 letter abbreviation");
+        return next();
+    }
+
+    if (req.body.uname.length > 50) {
+        res.status(400).send("Username is too long");
+    }
+
+    // eslint-disable-next-line
+    if (/[$&+,/:;=?@ "<>#%{}|\\^~\[\]`]/.test(req.body.uname)) {
+        res.status(400).send("Username cannot contain any special characters");
+        return next();
+    }
+
+    // eslint-disable-next-line
+    if (/[^\u0000-\u007f]/.test(req.body.uname)) {
+        res.status(400).send("Username cannot contain any Unicode characters");
+        return next();
+    }
+
+    if (req.body.createpin !== process.env.ACCOUNT_PIN) {
+        res.status(400).send("Incorrect Account Creation PIN");
+        return next();
+    }
+
+    if (req.body.pass1 !== req.body.pass2) {
+        res.status(400).send("Passwords do not match");
+        return next();
+    }
+
+    if (req.body.pass1.length < 8) {
+        res.status(400).send("Password length must be greater than 8 characters");
+        return next();
+    }
+
+    // force emails lowercase for OIDC purposes
+    req.body.email = req.body.email.toLowerCase();
+
+    bcrypt.hash(req.body.pass1, bcrypt_rounds, async(error, hash) => {
+        if (error) {
+            logger.error(error);
+            res.status(500).send("Internal Server Error");
+            return next();
+        }
+
+        try {
+            // First try and create the account
+            await Models.account.createUser(req.body, hash);
+
+            // Setup the return object
+            const user = {
+                uname: req.body.uname,
+            };
+
+            // Get all privilege levels
+            const [response] = await Models.account.getUserAccessLevel(req.body.uname);
+            if (response[0].maxPrivilege !== null) {
+                user.viewFinancials = true;
+                user.viewApprove = response[0].maxAmount > 0;
+                user.viewOfficer = response[0].maxPrivilege >= ACCESS_LEVEL.officer;
+                user.viewTreasurer = response[0].maxPrivilege >= ACCESS_LEVEL.treasurer;
+            } else {
+                user.viewFinancials = false;
+                user.viewApprove = false;
+                user.viewOfficer = false;
+                user.viewTreasurer = false;
+            }
+
+            // Generate the API key now
+            const response_1 = await Models.account.generateAPIKey(req.body.uname);
+            res.cookie("apikey", response_1, { maxAge:1000*60*60*24, sameSite:"strict",}); // cookie is valid for 24 hours
+            res.status(201).send(user);
+            return next();
+        } catch (err) {
+            // Username already exists
+            if (err.code === "ER_DUP_ENTRY") {
+                res.status(400).send("Username or email already exists");
+                return next();
+            } else {
+                logger.error(err.stack);
+                res.status(500).send("Internal Server Error");
+                return next();
+            }
+        }
+    });
+});
+
+/*
+    Logs user in
+*/
+router.post("/password", async(req, res, next) => {
     if (req.body.uname === undefined || req.body.uname === "" ||
         req.body.pass === undefined || req.body.pass === "") {
         res.status(400).send("Fill out login details");
